+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* BREAKING: Transform keys in Publishing API GraphQL responses to have underscore keys [PR](https://github.com/alphagov/gds-api-adapters/pull/1310).
 * Remove router-api [PR](https://github.com/alphagov/gds-api-adapters/pull/1308)
 
 ## 97.6.0